feat: introduced abstract filter class which different persistence implementations can use.
